utils: stop pulling frontend stack into zambda bundles
utils/lib/main.ts re-exported ./frontend via "export * from './frontend'",
which dragged @sentry/react, react, react-dom, react-router-dom, and
react-i18next into ~99.6% of zambda bundles. esbuild was already
tree-shaking the frontend code itself (utils/frontend contributed
~0.4 kB per bundle), but the bare module imports inside that subtree
kept the underlying packages retained because those packages don't
declare "sideEffects": false.

helpers/i18nextCreator.ts had the same problem: it lives under
helpers/ but imports react-i18next, so re-exporting it from the
helpers barrel pulled the React stack into every bundle via a
separate path.

Changes:
- Remove `export * from './frontend'` from utils/lib/main.ts.
- Move helpers/i18nextCreator.ts → frontend/i18nextCreator.ts and
  re-export it from the frontend barrel.
- Move parseCommaSeparatedTags out of frontend/configurations/sentry.ts
  into helpers/parseCommaSeparatedTags.ts so the (pure) string parser
  used by the zambda Sentry wrapper stays reachable from `utils`
  without dragging in @sentry/react.
- Move the test alongside the moved module.
- Update all 32 app call sites: frontend symbols (useErrorQuery,
  useSuccessQuery, useCookie, setupSentry, i18nextCreator) now
  import from 'utils/lib/frontend'. Backend-safe symbols keep
  importing from 'utils'.

After the change, react / react-dom / react-router-dom /
react-i18next / @sentry/react no longer appear in any zambda bundle.
