+++ apps/ehr/src/components/ScheduleInformation.tsx
@@ -23,7 +23,8 @@ import { DateTime } from 'luxon';
 import { enqueueSnackbar } from 'notistack';
 import { ReactElement, useMemo, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { APIError, isApiError, SchedulesAndOwnerListItem, useErrorQuery } from 'utils';
+import { APIError, isApiError, SchedulesAndOwnerListItem } from 'utils';
+import { useErrorQuery } from 'utils/lib/frontend';
 import { listScheduleOwners } from '../api/api';
 import { dataTestIds } from '../constants/data-test-ids';
 import { useApiClients } from '../hooks/useAppClients';
