+++ apps/ehr/src/features/visits/shared/stores/appointment/appointment.queries.ts
@@ -54,9 +54,8 @@ import {
   PromiseReturnType,
   ReviewAndSignData,
   UpdateMedicationOrderInput,
-  useErrorQuery,
-  useSuccessQuery,
 } from 'utils';
+import { useErrorQuery, useSuccessQuery } from 'utils/lib/frontend';
 import { OystehrTelemedAPIClient } from '../../api/oystehrApi';
 import { useOystehrAPIClient } from '../../hooks/useOystehrAPIClient';
 import { useAppointmentData } from './appointment.store';
