+++ apps/ehr/src/hooks/useGetPatient.ts
@@ -11,8 +11,8 @@ import {
   ORG_TYPE_PAYER_CODE,
   PromiseReturnType,
   RemoveCoverageZambdaInput,
-  useSuccessQuery,
 } from 'utils';
+import { useSuccessQuery } from 'utils/lib/frontend';
 import { OystehrTelemedAPIClient } from '../features/visits/shared/api/oystehrApi';
 import { getPatientNameSearchParams } from '../helpers/patientSearch';
 import { useApiClients } from './useAppClients';
