+++ apps/ehr/src/pages/SchedulePage.tsx
@@ -29,8 +29,8 @@ import {
   scheduleTypeFromFHIRType,
   TIMEZONES,
   UpdateScheduleParams,
-  useSuccessQuery,
 } from 'utils';
+import { useSuccessQuery } from 'utils/lib/frontend';
 import { createSchedule, getSchedule, updateSchedule } from '../api/api';
 import CustomBreadcrumbs from '../components/CustomBreadcrumbs';
 import Loading from '../components/Loading';
