+++ apps/ehr/src/rcm/features/claims-queue/ClaimsQueueFilters.tsx
@@ -4,7 +4,8 @@ import { useQuery } from '@tanstack/react-query';
 import { FC, useEffect, useMemo, useState } from 'react';
 import { useClaimsQueueStore, useGetFacilities, useGetInsurancePlans, useGetOrganizations } from 'src/rcm/state';
 import { useDebounce } from 'src/shared/hooks/useDebounce';
-import { FHIR_EXTENSION, GetEmployeesResponse, getSelectors, useSuccessQuery } from 'utils';
+import { FHIR_EXTENSION, GetEmployeesResponse, getSelectors } from 'utils';
+import { useSuccessQuery } from 'utils/lib/frontend';
 import { getEmployees } from '../../../api/api';
 import { useApiClients } from '../../../hooks/useAppClients';
 import { VirtualizedAutocomplete } from '../claim/modals/components';
