+++ apps/ehr/src/rcm/state/claim/claim.queries.ts
@@ -13,8 +13,8 @@ import {
   RelatedPerson,
 } from 'fhir/r4b';
 import { findResourceByType, generateOpByResourceData, getCoverageRelatedResources } from 'src/rcm/utils';
-import { useSuccessQuery } from 'utils';
 import { INSURANCE_PLAN_PAYER_META_TAG_CODE } from 'utils';
+import { useSuccessQuery } from 'utils/lib/frontend';
 import { useApiClients } from '../../../hooks/useAppClients';
 
 export const useGetClaim = (
