+++ apps/intake/src/pages/MyPatients.tsx
@@ -5,7 +5,8 @@ import { DateTime } from 'luxon';
 import { useMemo, useState } from 'react';
 import { FieldValues } from 'react-hook-form';
 import { generatePath, Outlet, useNavigate, useOutletContext, useParams } from 'react-router-dom';
-import { APIError, getPatientInfoFullName, isApiError, PatientInfo, useErrorQuery } from 'utils';
+import { APIError, getPatientInfoFullName, isApiError, PatientInfo } from 'utils';
+import { useErrorQuery } from 'utils/lib/frontend';
 import ottehrApi from '../api/ottehrApi';
 import { intakeFlowPageRoute } from '../App';
 import { PageContainer } from '../components';
