+++ packages/utils/lib/frontend/i18nextCreator.ts
