+++ packages/utils/lib/frontend/index.ts
@@ -1,2 +1,3 @@
 export * from './hooks';
 export * from './configurations';
+export * from './i18nextCreator';
