+++ packages/utils/lib/helpers/index.ts
@@ -3,8 +3,8 @@ export * from './check-office-open';
 export * from './create-demo-visits';
 export * from './deploy';
 export * from './helpers';
-export * from './i18nextCreator';
 export * from './in-house-labs';
+export * from './parseCommaSeparatedTags';
 export * from './labs';
 export * from './medications';
 export * from './operations';
